Gathers data into array
So far, only gathers data into an array in importer.php. It is now
needed to add a page chooser (for a different calendar) and for the
events to be exported to ICS. Also, a prompt for checking the data is
correct for the first (or more) records may be useful.